Set appropriate cache headers to expire in AgendasController
This is to correct a small edge-case. What would happen is:
- User ends up on any given /edit
- User adds a new agenda course
- User clicks submit
- On the /schedules page, the user clicks back in the browser
- User then re-submits the form

What would end up happening is the agenda was trying to create the new agenda_course, with the same index, as the page cache had not expired. As such, we then got a NotUnique error on the agenda_course.